feat: Add automatic keyword listening upon GUI launch
- Created `listenForKeyword` function to continuously listen for a keyword until detected
- Created `startListeningThread` function to run `listenForKeyword` in a separate thread

The `listenForKeyword` function continuously records audio and transcribes it using the ASR module. If the keyword "sherpa" is detected, it triggers a recording which will be used to run a command. The `startListeningThread` function creates a separate thread for the `listenForKeyword` function, allowing it to run concurrently with the GUI.

The user experience has been enhanced by eliminating the need to press a button, enabling voice-based control when the GUI is launched.