feat: Add function to search for documents based on user input
This commit introduces a new function that automates the process of searching for documents. The function moves the mouse cursor to the Windows search bar located in the bottom left corner and clicks on it.
If the last word of the input string is "document", the user is prompted to enter the name of the document they want to search for. Otherwise, the function extracts the document name from the input string by finding the position of the word "document" and concatenating all the words that come after it.
The function then types the document name into the search bar and presses the Enter key to initiate the search.