feat: Add command-line option to disable keyword listening
Introduce a new command-line option `-d` or `--disable` to disable keyword listening functionality. When the option is provided, the program will not start the keyword listening thread.
The change is implemented by integrating the `argparse` library to parse command-line arguments and determine whether keyword listening should be enabled or disabled.